remove cseries_strings file
- move static string classes to static_arrays.h
- move string helper functions to cseries
- fix c_static_wchar_string print member function to use T for length
- add print member for regular static string
