+++ xlive/Blam/Engine/cseries/cseries_windows_minidump.h
@@ -1,4 +1,4 @@
 #pragma once
-#include "cseries_strings.h"
+
 
 void write_crash_dump_files(_EXCEPTION_POINTERS* ExceptionInfo, c_static_wchar_string<MAX_PATH>* report_path, c_static_wchar_string<MAX_PATH>* archive_path);
