+++ xlive/Blam/Engine/cseries/cseries_windows_minidump_logs.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "cseries_strings.h"
+
 
 /* enums */
