+++ xlive/Blam/Engine/main/main_game.cpp
@@ -2,7 +2,7 @@
 #include "main_game.h"
 
 #include "cache/cache_files.h"
-#include "cseries/cseries_strings.h"
+
 #include "game/game.h"
 #include "saved_games/game_variant.h"
