+++ xlive/Blam/Engine/render/render_cartographer_ingame_ui.cpp
@@ -2,7 +2,7 @@
 #include "render_cartographer_ingame_ui.h"
 
 #include "cartographer/twizzler/twizzler.h"
-#include "cseries/cseries_strings.h"
+
 #include "game/game.h"
 #include "game/game_engine.h"
 #include "main/main_game_time.h"
