+++ xlive/Blam/Engine/tag_files/tag_loader/xml/xml_definition_block.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "tinyxml/tinyxml2.h"
 
-#include "cseries/cseries_strings.h"
+
 #include "tag_files/tag_loader/tag_injection_define.h"
