+++ xlive/H2MOD/GUI/imgui_integration/MOTD.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "imgui_handler.h"
 
-#include "cseries/cseries_strings.h"
+
 #include "game/player_constants.h"
 #include "input/input_abstraction.h"
 #include "input/input_windows.h"
