+++ xlive/H2MOD/Modules/Achievements/Achievements.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "Achievements.h"
 
-#include "cseries/cseries_strings.h"
+
 #include "H2MOD/Modules/Accounts/Accounts.h"
 #include "H2MOD/Modules/Shell/Config.h"
 #include "rapidjson/document.h"
