+++ xlive/H2MOD/Modules/Shell/Startup/Startup.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "cseries/cseries_strings.h"
+
 
 void InitH2Startup();
 void H2DedicatedServerStartup();
