+++ xlive/XLive/XUser/XUserContext.cpp
@@ -2,7 +2,7 @@
 #include "XUserContext.h"
 
 #include "cartographer/discord/discord_interface.h"
-#include "cseries/cseries_strings.h"
+
 #include "game/game.h"
 #include "networking/Session/network_session.h"
 #include "text/unicode.h"
